VACMS-20046 Move widget types to platform folder to remove cross-app dependency
