+++ src/applications/static-pages/BTSSS-login/createBTSSSLogin.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import widgetTypes from '../widgetTypes';
+import widgetTypes from 'platform/site-wide/widgetTypes';
 
 export default store => {
   const root = document.querySelector(
