+++ src/applications/static-pages/facilities/createFacilityPageSatelliteLocations.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
+import widgetTypes from 'platform/site-wide/widgetTypes';
 import { fetchMultiFacility } from './actions';
-import widgetTypes from '../widgetTypes';
 
 export default async function createFacilityPageSatelliteLocations(store) {
   let facilityID = '';
