+++ src/applications/static-pages/facilities/vet-center/createNearByVALocations.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-
-import widgetTypes from '../../widgetTypes';
+import widgetTypes from 'platform/site-wide/widgetTypes';
 
 export default async function createNearByVALocations(store) {
   const nearByVALocationsWidget = document.querySelector(
