+++ src/platform/site-wide/banners/index.js
@@ -1,7 +1,5 @@
-// Node modules.
 import React from 'react';
-// Relative imports.
-import widgetTypes from 'applications/static-pages/widgetTypes';
+import widgetTypes from 'platform/site-wide/widgetTypes';
 import startReactApp from '../../startup/react';
 
 // Are you looking for where this is used?
