+++ src/platform/site-wide/widgetTypes.js
